Move the inmemory provider to a new mod
